+++ rust/aura/src/utils.rs
@@ -2,6 +2,7 @@
 
 use crate::error::Error;
 use rustyline::Editor;
+use std::process::Command;
 
 // TODO Localize the acceptance chars.
 /// Prompt the user for confirmation.
@@ -11,5 +12,14 @@ pub(crate) fn prompt(msg: &str) -> Result<(), Error> {
         Ok(line) if line.is_empty() || line == "y" || line == "Y" => Ok(()),
         Ok(_) => Err(Error::Rejected),
         Err(e) => Err(Error::RustyLine(e)),
+    }
+}
+
+/// Make a shell call to `pacman`.
+pub(crate) fn pacman(args: Vec<String>) -> Result<(), Error> {
+    match Command::new("pacman").args(args).status() {
+        Err(e) => Err(Error::IO(e)),
+        Ok(es) if es.success() => Ok(()),
+        Ok(_) => Err(Error::PacmanError),
     }
 }